requirements.txt
python-telegram-bot==20.7
alibabacloud-bssopenapi20171214==5.0.0
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.11
alibabacloud-endpoint-util==0.0.3
cryptography==41.0.7
aiofiles==23.2.1
aiohttp==3.9.1
python-dotenv==1.0.0
